+++ crates/bacnet-transport/src/bip/tests.rs
@@ -1,6 +1,4 @@
 use super::*;
-use bacnet_encoding::npdu::{decode_npdu, encode_npdu, Npdu, NpduAddress};
-use bacnet_types::MacAddr;
 use bytes::{Bytes, BytesMut};
 use tokio::time::{timeout, Duration};
 
@@ -783,114 +781,6 @@ async fn socket_is_broadcast_capable_and_binds_inaddr_any() {
     );
 
     transport.stop().await.unwrap();
-}
-
-#[tokio::test]
-async fn original_broadcast_npdu_preserves_global_npdu_destination() {
-    let sink = UdpSocket::bind(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0))
-        .await
-        .unwrap();
-    let sink_port = sink.local_addr().unwrap().port();
-    let source_socket = UdpSocket::bind(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0))
-        .await
-        .unwrap();
-    let mut transport = BipTransport::new(Ipv4Addr::LOCALHOST, sink_port, Ipv4Addr::LOCALHOST);
-    transport.socket = Some(Arc::new(source_socket));
-
-    let mut npdu_buf = BytesMut::new();
-    encode_npdu(
-        &mut npdu_buf,
-        &Npdu {
-            destination: Some(NpduAddress {
-                network: 0xFFFF,
-                mac_address: MacAddr::new(),
-            }),
-            hop_count: 255,
-            payload: Bytes::from_static(&[0x10, 0x08]),
-            ..Default::default()
-        },
-    )
-    .unwrap();
-    let npdu = npdu_buf.freeze();
-
-    transport.send_broadcast(&npdu).await.unwrap();
-
-    let mut recv_buf = [0u8; 2048];
-    let (len, _addr) = timeout(Duration::from_secs(2), sink.recv_from(&mut recv_buf))
-        .await
-        .expect("timed out waiting for Original-Broadcast-NPDU")
-        .unwrap();
-    let frame = decode_bvll(&recv_buf[..len]).unwrap();
-
-    assert_eq!(frame.function, BvlcFunction::ORIGINAL_BROADCAST_NPDU);
-    assert_eq!(
-        frame.payload.as_ref(),
-        npdu.as_ref(),
-        "Original-Broadcast-NPDU must preserve remote/global NPDU addressing bytes"
-    );
-
-    let decoded = decode_npdu(frame.payload).unwrap();
-    let destination = decoded
-        .destination
-        .expect("global broadcast NPDU must keep DNET/DADR destination");
-    assert_eq!(destination.network, 0xFFFF);
-    assert!(destination.mac_address.is_empty());
-    assert_eq!(decoded.hop_count, 255);
-    assert_eq!(decoded.payload.as_ref(), &[0x10, 0x08]);
-}
-
-#[tokio::test]
-async fn original_broadcast_npdu_preserves_remote_npdu_destination() {
-    let sink = UdpSocket::bind(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0))
-        .await
-        .unwrap();
-    let sink_port = sink.local_addr().unwrap().port();
-    let source_socket = UdpSocket::bind(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 0))
-        .await
-        .unwrap();
-    let mut transport = BipTransport::new(Ipv4Addr::LOCALHOST, sink_port, Ipv4Addr::LOCALHOST);
-    transport.socket = Some(Arc::new(source_socket));
-
-    let mut npdu_buf = BytesMut::new();
-    encode_npdu(
-        &mut npdu_buf,
-        &Npdu {
-            destination: Some(NpduAddress {
-                network: 0x1234,
-                mac_address: MacAddr::new(),
-            }),
-            hop_count: 64,
-            payload: Bytes::from_static(&[0x10, 0x09]),
-            ..Default::default()
-        },
-    )
-    .unwrap();
-    let npdu = npdu_buf.freeze();
-
-    transport.send_broadcast(&npdu).await.unwrap();
-
-    let mut recv_buf = [0u8; 2048];
-    let (len, _addr) = timeout(Duration::from_secs(2), sink.recv_from(&mut recv_buf))
-        .await
-        .expect("timed out waiting for Original-Broadcast-NPDU")
-        .unwrap();
-    let frame = decode_bvll(&recv_buf[..len]).unwrap();
-
-    assert_eq!(frame.function, BvlcFunction::ORIGINAL_BROADCAST_NPDU);
-    assert_eq!(
-        frame.payload.as_ref(),
-        npdu.as_ref(),
-        "Original-Broadcast-NPDU must preserve remote NPDU addressing bytes"
-    );
-
-    let decoded = decode_npdu(frame.payload).unwrap();
-    let destination = decoded
-        .destination
-        .expect("remote broadcast NPDU must keep DNET/DADR destination");
-    assert_eq!(destination.network, 0x1234);
-    assert!(destination.mac_address.is_empty());
-    assert_eq!(decoded.hop_count, 64);
-    assert_eq!(decoded.payload.as_ref(), &[0x10, 0x09]);
 }
 
 #[tokio::test]
